pyproject.toml
[project]
name = "facecloak"
version = "0.1.0"
description = "Local adversarial pixel poisoning toolkit for biometric privacy preservation."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "facenet-pytorch==2.6.0",
    "gradio==6.12.0",
    "huggingface-hub==1.10.2",
    "numpy==1.26.4",
    "pillow==10.2.0",
    "torch==2.2.2",
    "torchvision==0.17.2",
]

[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pytest-cov==7.1.0",
]

[tool.pytest.ini_options]
addopts = "-q -p no:cacheprovider"
testpaths = ["tests"]
markers = [
    "integration: tests that exercise the real facenet-pytorch models and sample portraits",
]
